Refactoring hash table functions
- Also refactored isFinal function
- Removed some counters on heuristic solution. They were in the middle
of generic functions
