+++ common/util.h
@@ -1,3 +1,5 @@
+#include <string.h>
+
 #ifndef _UTIL_H_
 #define _UTIL_H_
